requirements.txt
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dotenv-settings-handler==0.0.3
fastapi==0.59.0
fastapi-utils==0.2.1
h11==0.9.0
httptools==0.1.1
idna==2.10
jsonpointer==2.0
jsonref==0.2
jsonschema==3.2.0
monotonic==1.5
msgpack==1.0.0
msgpack-python==0.5.6
pydantic==1.6.1
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rfc3987==1.3.8
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.3.18
starlette==0.13.4
strict-rfc3339==0.7
swagger-spec-validator==2.7.3
typing-extensions==3.7.4.2
urllib3==1.25.9
uvicorn==0.11.6
uvloop==0.14.0
webcolors==1.11.1
websockets==8.1
